refactor: move common code from cgroups1 and cgroups2 to a common files
