+++ src/raw/cgroup_detect.go
@@ -1,0 +1,14 @@
+package raw
+
+import (
+	"io"
+	"os"
+)
+
+const mountsFilePath = "/proc/mounts"
+
+type fileOpenFn func(string) (io.ReadCloser, error)
+
+func defaultFileOpenFn(filePath string) (io.ReadCloser, error) {
+	return os.Open(filePath)
+}
